requirements.txt
jwt==1.3.1
passlib==1.7.4
SQLAlchemy==1.4.22
fastapi==0.70.0
python-dotenv==0.19.1
PyJWT==2.3.0